+++ src/main/java/math/Operable.java
@@ -1,10 +1,11 @@
 package math;
 
+import java.math.BigDecimal;
 import modelo.Enterprise;
 import modelo.Indicator;
 
 public interface Operable {
-	public double reduce(Enterprise enterprise, int year);
+	public BigDecimal reduce(Enterprise enterprise, int year);
 	
 	public Boolean includes(Indicator indicator);
 }
